Make all our zip operations Robust (BL-9926)
